requirements.txt
# Add your dependencies here
# For example:
# requests==2.28.1
python-dotenv==1.0.0